REF: Move `app_loading`
