+++ octue/twined/app_loading.py
@@ -3,7 +3,6 @@ import logging
 import os
 import sys
 
-
 logger = logging.getLogger(__name__)
