+++ octue/twined/runner.py
@@ -13,11 +13,11 @@ from jsonschema import ValidationError
 from jsonschema import validate as jsonschema_validate
 
 from octue import exceptions
-from octue.app_loading import AppFrom
 from octue.diagnostics import Diagnostics
 from octue.log_handlers import AnalysisLogFormatterSwitcher
 from octue.resources.datafile import downloaded_files
 from octue.twined import Twine
+from octue.twined.app_loading import AppFrom
 import octue.twined.exceptions
 from octue.twined.resources import Child
 from octue.twined.resources.analysis import CLASS_MAP, Analysis
